binary_port: Clean up tests for enabled/disabled binary port functions
